requirements.txt
tqdm
numpy
matplotlib
pandas
jupyter
opencv-python
rasterio
albumentations
torch -f https://download.pytorch.org/whl/nightly/cu110/torch_nightly.html
torchvision==0.5.0
